@@ -1,17 +1,13 @@
 package com.mrbysco.woolytrees.registry;
 
-import com.google.common.collect.ImmutableList;
 import com.mrbysco.woolytrees.Reference;
 import com.mrbysco.woolytrees.trees.features.FancyWoolPlacer;
-import net.minecraft.core.Registry;
-import net.minecraft.data.BuiltinRegistries;
-import net.minecraft.data.worldgen.Features;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.data.worldgen.features.FeatureUtils;
+import net.minecraft.data.worldgen.features.TreeFeatures;
 import net.minecraft.util.valueproviders.ConstantInt;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
-import net.minecraft.world.level.levelgen.feature.configurations.FeatureConfiguration;
 import net.minecraft.world.level.levelgen.feature.configurations.TreeConfiguration;
 import net.minecraft.world.level.levelgen.feature.featuresize.TwoLayersFeatureSize;
 import net.minecraft.world.level.levelgen.feature.foliageplacers.BlobFoliagePlacer;
@@ -19,11 +15,16 @@ import net.minecraft.world.level.levelgen.feature.foliageplacers.FancyFoliagePla
 import net.minecraft.world.level.levelgen.feature.stateproviders.SimpleStateProvider;
 import net.minecraft.world.level.levelgen.feature.trunkplacers.FancyTrunkPlacer;
 import net.minecraft.world.level.levelgen.feature.trunkplacers.StraightTrunkPlacer;
+import net.minecraft.world.level.levelgen.feature.trunkplacers.TrunkPlacerType;
 
+import java.util.List;
 import java.util.OptionalInt;
 import java.util.Random;
 
 public class WoolyFeatureConfig {
+    public static final TrunkPlacerType<FancyTrunkPlacer> STRAIGHT_FANCY_TRUNK_PLACER = TrunkPlacerType.register(
+            Reference.MOD_PREFIX + "straight_fancy_trunk_placer", FancyTrunkPlacer.CODEC);
+
     public static final Random rand = new Random();
 
     private static final BlockState WHITE_WOOL_LOG = Blocks.WHITE_WOOL.defaultBlockState();
@@ -31,32 +32,72 @@ public class WoolyFeatureConfig {
     private static final BlockState WOOLY_SAPLING = WoolyRegistry.WOOLY_SAPLING.get().defaultBlockState();
     private static final BlockState JEB_SAPLING = WoolyRegistry.JEB_SAPLING.get().defaultBlockState();
 
-    public static final ConfiguredFeature<TreeConfiguration, ?> FANCY_WOOL = register("fancy_wool", WoolyRegistry.WOOLY_TREE.get().configured(
-            (new TreeConfiguration.TreeConfigurationBuilder(
-                    new SimpleStateProvider(WHITE_WOOL_LOG),
-                    new FancyTrunkPlacer(3, 11, 0),
-                    new SimpleStateProvider(GREEN_WOOL_LEAVES),
-                    new SimpleStateProvider(WOOLY_SAPLING),
-                    new FancyFoliagePlacer(ConstantInt.of(2), ConstantInt.of(4), 4),
-                    new TwoLayersFeatureSize(0, 0, 0, OptionalInt.of(4)))).ignoreVines().build()));
+    public static final ConfiguredFeature<TreeConfiguration, ?> FANCY_WOOL = FeatureUtils.register("fancy_wool", WoolyRegistry.WOOLY_TREE.get()
+            .configured(createFancyWool().build()));
 
-    public static final ConfiguredFeature<TreeConfiguration, ?> FANCY_WOOL_BEES_005 = register("fancy_wool_with_bees_005", WoolyRegistry.WOOLY_TREE.get().configured(
-            FANCY_WOOL.config().withDecorators(ImmutableList.of(Features.Decorators.BEEHIVE_005))));
+    public static final ConfiguredFeature<TreeConfiguration, ?> FANCY_WOOL_BEES_005 = FeatureUtils.register("fancy_wool_with_bees_005", WoolyRegistry.WOOLY_TREE.get()
+            .configured(createFancyWool().decorators(List.of(TreeFeatures.BEEHIVE_005)).build()));
 
-    public static final ConfiguredFeature<TreeConfiguration, ?> WOOL = register("wool", WoolyRegistry.WOOLY_TREE.get().configured(
-            (new TreeConfiguration.TreeConfigurationBuilder(
-                    new SimpleStateProvider(WHITE_WOOL_LOG),
-                    new StraightTrunkPlacer(4, 2, 0),
-                    new SimpleStateProvider(GREEN_WOOL_LEAVES),
-                    new SimpleStateProvider(WOOLY_SAPLING),
-                    new BlobFoliagePlacer(ConstantInt.of(2), ConstantInt.of(0), 3),
-                    new TwoLayersFeatureSize(1, 0, 1))).ignoreVines().build()));
+    public static final ConfiguredFeature<TreeConfiguration, ?> WOOL = FeatureUtils.register("wool", WoolyRegistry.WOOLY_TREE.get().configured(
+            createWool().build()));
+
+    public static final ConfiguredFeature<TreeConfiguration, ?> WOOL_BEES_005 = FeatureUtils.register("wool_with_bees_005", WoolyRegistry.WOOLY_TREE.get()
+            .configured(createWool().decorators(List.of(TreeFeatures.BEEHIVE_005)).build()));
+
+    private static TreeConfiguration.TreeConfigurationBuilder createFancyWool() {
+        return (new TreeConfiguration.TreeConfigurationBuilder(
+                SimpleStateProvider.simple(WHITE_WOOL_LOG),
+                new FancyWoolPlacer(3, 11, 0),
+                SimpleStateProvider.simple(GREEN_WOOL_LEAVES),
+                new FancyFoliagePlacer(ConstantInt.of(2), ConstantInt.of(4), 4),
+                new TwoLayersFeatureSize(0, 0, 0, OptionalInt.of(4)))).ignoreVines();
+    }
+
+    private static TreeConfiguration.TreeConfigurationBuilder createWool() {
+        return (new TreeConfiguration.TreeConfigurationBuilder(
+                SimpleStateProvider.simple(WHITE_WOOL_LOG),
+                new StraightTrunkPlacer(4, 2, 0),
+                SimpleStateProvider.simple(GREEN_WOOL_LEAVES),
+                new BlobFoliagePlacer(ConstantInt.of(2), ConstantInt.of(0), 3),
+                new TwoLayersFeatureSize(1, 0, 1))).ignoreVines();
+    }
 
 
-    public static final ConfiguredFeature<TreeConfiguration, ?> WOOL_BEES_005 = register("fancy_wool_with_bees_005", WoolyRegistry.WOOLY_TREE.get().configured(
-            WOOL.config().withDecorators(ImmutableList.of(Features.Decorators.BEEHIVE_005))));
+    public static TreeConfiguration.TreeConfigurationBuilder getFancyJeb() {
+        return (new TreeConfiguration.TreeConfigurationBuilder(
+                SimpleStateProvider.simple(getRandomLog()),
+                new FancyWoolPlacer(3, 11, 0),
+                SimpleStateProvider.simple(getRandomLeaves()),
+                new FancyFoliagePlacer(ConstantInt.of(2), ConstantInt.of(4), 4),
+                new TwoLayersFeatureSize(0, 0, 0, OptionalInt.of(4)))).ignoreVines();
+    }
 
-    public static BlockState getRandomleaves() {
+    public static TreeConfiguration getFancyJebConfiguration() {
+        return getFancyJeb().build();
+    }
+
+    public static TreeConfiguration getFancyJebWithBeehives() {
+        return getFancyJeb().decorators(List.of(TreeFeatures.BEEHIVE_005)).build();
+    }
+
+    public static TreeConfiguration getJebConfiguration() {
+        return getJeb().build();
+    }
+
+    public static TreeConfiguration getJebWithBeehives() {
+        return getJeb().decorators(List.of(TreeFeatures.BEEHIVE_005)).build();
+    }
+
+    public static TreeConfiguration.TreeConfigurationBuilder getJeb() {
+        return (new TreeConfiguration.TreeConfigurationBuilder(
+                SimpleStateProvider.simple(getRandomLog()),
+                new StraightTrunkPlacer(4, 2, 0),
+                SimpleStateProvider.simple(getRandomLeaves()),
+                new BlobFoliagePlacer(ConstantInt.of(2), ConstantInt.of(0), 3),
+                new TwoLayersFeatureSize(1, 0, 1))).ignoreVines();
+    }
+
+    public static BlockState getRandomLeaves() {
         BlockState randomState = GREEN_WOOL_LEAVES;
         if(!WoolyTags.WOOLY_LEAVES.getValues().isEmpty()) {
             randomState = WoolyTags.WOOLY_LEAVES.getRandomElement(rand).defaultBlockState();
@@ -70,36 +111,5 @@ public class WoolyFeatureConfig {
             randomState = WoolyTags.WOOLY_LOGS.getRandomElement(rand).defaultBlockState();
         }
         return randomState;
-    }
-
-    public static TreeConfiguration getFancyJeb() {
-        return getFancyJebWithBeehives().withDecorators(ImmutableList.of(Features.Decorators.BEEHIVE_005));
-    }
-
-    public static TreeConfiguration getFancyJebWithBeehives() {
-        return (new TreeConfiguration.TreeConfigurationBuilder(
-                new SimpleStateProvider(getRandomLog()),
-                new FancyWoolPlacer(3, 11, 0),
-                new SimpleStateProvider(getRandomleaves()),
-                new SimpleStateProvider(JEB_SAPLING),
-                new FancyFoliagePlacer(ConstantInt.of(2), ConstantInt.of(4), 4),
-                new TwoLayersFeatureSize(0, 0, 0, OptionalInt.of(4)))).ignoreVines().build();
-    }
-
-    public static TreeConfiguration getJebWithBeehives() {
-        return getJeb().withDecorators(ImmutableList.of(Features.Decorators.BEEHIVE_005));
-    }
-    public static TreeConfiguration getJeb() {
-        return (new TreeConfiguration.TreeConfigurationBuilder
-                (new SimpleStateProvider(getRandomLog()),
-                        new StraightTrunkPlacer(4, 2, 0),
-                        new SimpleStateProvider(getRandomleaves()),
-                        new SimpleStateProvider(JEB_SAPLING),
-                        new BlobFoliagePlacer(ConstantInt.of(2), ConstantInt.of(0), 3),
-                        new TwoLayersFeatureSize(1, 0, 1))).ignoreVines().build();
-    }
-
-    private static <FC extends FeatureConfiguration> ConfiguredFeature<FC, ?> register(String key, ConfiguredFeature<FC, ?> feature) {
-        return Registry.register(BuiltinRegistries.CONFIGURED_FEATURE, new ResourceLocation(Reference.MOD_ID, key), feature);
     }
 }
